better digest email when a new branch is pushed
In digest mode (git-notify -n 0) -- where only one email is sent summarizing all the pushed commits -- the email followed a different format that only described the latest commit on a new branch. This was misleading.

Now, when a new ref is pushed in digest mode, we backtrack to the most recent named ancestor commit and generate the same summary email for that range of commits. Which should be the commits that have occurred on this new branch since it diverged.